Impr(general): Finish namespace renaming
