+++ skbc/AST/IfNode.cs
@@ -1,5 +1,4 @@
-using Data.ErrorHandling;
-using Data.Objects;
+using Compiler.Data;
 using System.Reflection.Emit;
 
 namespace Compiler.AST;
