+++ skbc/AST/RecordNode.cs
@@ -1,5 +1,4 @@
-using Data.ErrorHandling;
-using Data.Objects;
+using Compiler.Data;
 
 namespace Compiler.AST;
 public class RecordNode : Node {
