+++ skbc/Data/Objects.cs
@@ -1,225 +1,223 @@
-namespace Data.Objects {
-	public struct Position {
-		public Position(int x, int y) {
-			this.row = x;
-			this.col = y;
-		}
-		
-		private int row {get; set;}
-		private int col {get; set;}
+namespace Compiler.Data;
+public struct Position {
+	public Position(int x, int y) {
+		this.row = x;
+		this.col = y;
+	}
+	
+	private int row {get; set;}
+	private int col {get; set;}
 
-		public int Row() => row;
-		public int Col() => col;
+	public int Row() => row;
+	public int Col() => col;
 
-		public void NextLine() {
-			this.row += 1;
-			this.col = 0;
-		}
-		public void NextChar() {
-			this.col += 1;
-		}
-
-		public override string ToString() =>
-			$"({this.row},{this.col})";
+	public void NextLine() {
+		this.row += 1;
+		this.col = 0;
+	}
+	public void NextChar() {
+		this.col += 1;
 	}
 
-	public enum TokenCode {
-		builtin_type,
-		constant_value,
-		logic_op,
-		relation_op,
-		factor_op,
-		term_op,
-		identifier,
+	public override string ToString() =>
+		$"({this.row},{this.col})";
+}
 
-		// User-defined types
-		record_declaration,
-		array_declaration,
-		// Variable declaration
-		variable_declaration,
-		type_declaration,
-		// Assingnment
-		type_assignment,
-		bare_assignment,
-		is_assignment,
-		// Separators
-		dot,
-		comma,
-		semicolon,
-		// Delimeters
-		left_parenthesis,
-		right_parenthesis,
-		left_bracket,
-		right_bracket,
-		// Branching
-		if_statement,
-		then_statement,
-		else_statement,
-		// Loops
-		while_statement,
-		for_statement,
-		in_statement,
-		reverse_statement,
-		loop_start,
-		range_sign,
-		// Routines
-		print_routine,
-		routine_declaration,
-		one_line_body,
-		end_of_body,
-		// Sequence breaking
-		return_statement,
-		break_statement,
-		continue_statement,
-		end_of_file,
-		error,
-	}
+public enum TokenCode {
+	builtin_type,
+	constant_value,
+	logic_op,
+	relation_op,
+	factor_op,
+	term_op,
+	identifier,
 
-	public static class Precedence {
-		private static Dictionary<TokenCode, int> _precedence =
-			new Dictionary<TokenCode, int>() {
-				{TokenCode.dot				, 0},
-				{TokenCode.left_bracket		, 0},
-				{TokenCode.factor_op		, 1},
-				{TokenCode.term_op			, 2},
-				{TokenCode.relation_op		, 3},
-				{TokenCode.logic_op			, 4},
-				{TokenCode.left_parenthesis	, 5},
-			};
+	// User-defined types
+	record_declaration,
+	array_declaration,
+	// Variable declaration
+	variable_declaration,
+	type_declaration,
+	// Assingnment
+	type_assignment,
+	bare_assignment,
+	is_assignment,
+	// Separators
+	dot,
+	comma,
+	semicolon,
+	// Delimeters
+	left_parenthesis,
+	right_parenthesis,
+	left_bracket,
+	right_bracket,
+	// Branching
+	if_statement,
+	then_statement,
+	else_statement,
+	// Loops
+	while_statement,
+	for_statement,
+	in_statement,
+	reverse_statement,
+	loop_start,
+	range_sign,
+	// Routines
+	print_routine,
+	routine_declaration,
+	one_line_body,
+	end_of_body,
+	// Sequence breaking
+	return_statement,
+	break_statement,
+	continue_statement,
+	end_of_file,
+	error,
+}
 
-		public static int Order(TokenCode key) => _precedence[key];
-	}
+public static class Precedence {
+	private static Dictionary<TokenCode, int> _precedence =
+		new Dictionary<TokenCode, int>() {
+			{TokenCode.dot				, 0},
+			{TokenCode.left_bracket		, 0},
+			{TokenCode.factor_op		, 1},
+			{TokenCode.term_op			, 2},
+			{TokenCode.relation_op		, 3},
+			{TokenCode.logic_op			, 4},
+			{TokenCode.left_parenthesis	, 5},
+		};
 
-	public static class DedicatedWords {
-		private static List<string> _baseTypes =
-			new List<string>() { "integer", "real", "boolean", "void", "record", "array" };
-		private static Dictionary<string,TokenCode> _dedicatedWords =
-			new Dictionary<string,TokenCode>() {
-				{"integer",		TokenCode.builtin_type},
-				{"real",		TokenCode.builtin_type},
-				{"boolean", 	TokenCode.builtin_type},
-				{"record",		TokenCode.record_declaration},
-				{"array",		TokenCode.array_declaration},
-				{"and",			TokenCode.logic_op},
-				{"or",			TokenCode.logic_op},
-				{"xor",			TokenCode.logic_op},
-				{"not",			TokenCode.logic_op},
-				{"true",		TokenCode.constant_value},
-				{"false",		TokenCode.constant_value},
-				{"var",			TokenCode.variable_declaration},
-				{"is",			TokenCode.is_assignment},
-				{"type",		TokenCode.type_declaration},
-				{"end",			TokenCode.end_of_body},
-				{"if",			TokenCode.if_statement},
-				{"then",		TokenCode.then_statement},
-				{"else",		TokenCode.else_statement},
-				{"loop",		TokenCode.loop_start},
-				{"while",		TokenCode.while_statement},
-				{"for",			TokenCode.for_statement},
-				{"in",			TokenCode.in_statement},
-				{"reverse",		TokenCode.reverse_statement},
-				{"print",		TokenCode.print_routine},
-				{"routine", 	TokenCode.routine_declaration},
-				{"return",		TokenCode.return_statement},
-				{"break",		TokenCode.break_statement},
-				{"continue",	TokenCode.continue_statement}
-			};
+	public static int Order(TokenCode key) => _precedence[key];
+}
 
-		public static bool Keys(string key) =>
-			_dedicatedWords.ContainsKey(key);
+public static class DedicatedWords {
+	private static List<string> _baseTypes =
+		new List<string>() { "integer", "real", "boolean", "void", "record", "array" };
+	private static Dictionary<string,TokenCode> _dedicatedWords =
+		new Dictionary<string,TokenCode>() {
+			{"integer",		TokenCode.builtin_type},
+			{"real",		TokenCode.builtin_type},
+			{"boolean", 	TokenCode.builtin_type},
+			{"record",		TokenCode.record_declaration},
+			{"array",		TokenCode.array_declaration},
+			{"and",			TokenCode.logic_op},
+			{"or",			TokenCode.logic_op},
+			{"xor",			TokenCode.logic_op},
+			{"not",			TokenCode.logic_op},
+			{"true",		TokenCode.constant_value},
+			{"false",		TokenCode.constant_value},
+			{"var",			TokenCode.variable_declaration},
+			{"is",			TokenCode.is_assignment},
+			{"type",		TokenCode.type_declaration},
+			{"end",			TokenCode.end_of_body},
+			{"if",			TokenCode.if_statement},
+			{"then",		TokenCode.then_statement},
+			{"else",		TokenCode.else_statement},
+			{"loop",		TokenCode.loop_start},
+			{"while",		TokenCode.while_statement},
+			{"for",			TokenCode.for_statement},
+			{"in",			TokenCode.in_statement},
+			{"reverse",		TokenCode.reverse_statement},
+			{"print",		TokenCode.print_routine},
+			{"routine", 	TokenCode.routine_declaration},
+			{"return",		TokenCode.return_statement},
+			{"break",		TokenCode.break_statement},
+			{"continue",	TokenCode.continue_statement}
+		};
 
-		public static bool Vals(TokenCode code) =>
-			_dedicatedWords.ContainsValue(code);
+	public static bool Keys(string key) =>
+		_dedicatedWords.ContainsKey(key);
 
-		public static TokenCode Code(string key) {
-			if(!DedicatedWords.Keys(key)) return TokenCode.error;
-			return _dedicatedWords[key];
-			}
+	public static bool Vals(TokenCode code) =>
+		_dedicatedWords.ContainsValue(code);
 
-		public static bool BuiltInStrict(string type) =>
-			DedicatedWords.Code(type) == TokenCode.builtin_type;
-
-		public static bool BuiltIn(string type) =>
-			_baseTypes.Contains(type);
-	}
-
-	public struct ReturningStatus {
-		public bool returned;
-		public string ret_type;
-
-		public ReturningStatus(bool ret, string type) {
-			this.returned = ret;
-			this.ret_type = type;
+	public static TokenCode Code(string key) {
+		if(!DedicatedWords.Keys(key)) return TokenCode.error;
+		return _dedicatedWords[key];
 		}
 
-		public static ReturningStatus Copy(ReturningStatus stat) =>
-			new ReturningStatus(false, stat.ret_type);
+	public static bool BuiltInStrict(string type) =>
+		DedicatedWords.Code(type) == TokenCode.builtin_type;
+
+	public static bool BuiltIn(string type) =>
+		_baseTypes.Contains(type);
+}
+
+public struct ReturningStatus {
+	public bool returned;
+	public string ret_type;
+
+	public ReturningStatus(bool ret, string type) {
+		this.returned = ret;
+		this.ret_type = type;
 	}
 
-	public static class Returning {
-		private static Stack<ReturningStatus> returning =
-			new Stack<ReturningStatus>();
+	public static ReturningStatus Copy(ReturningStatus stat) =>
+		new ReturningStatus(false, stat.ret_type);
+}
 
-		public static void Push(ReturningStatus ret) =>
-			returning.Push(ret);
+public static class Returning {
+	private static Stack<ReturningStatus> returning =
+		new Stack<ReturningStatus>();
 
-		public static ReturningStatus Pop() =>
-			returning.Pop();
+	public static void Push(ReturningStatus ret) =>
+		returning.Push(ret);
 
-		public static ReturningStatus Peek() =>
-			returning.Peek();
+	public static ReturningStatus Pop() =>
+		returning.Pop();
 
-		public static int Count() =>
-			returning.Count();
+	public static ReturningStatus Peek() =>
+		returning.Peek();
+
+	public static int Count() =>
+		returning.Count();
+}
+
+public static class SeparatorList {
+	private static Dictionary<char, TokenCode> separatorCodes =
+		new Dictionary<char, TokenCode>() {
+			{',', TokenCode.comma},
+			{'(', TokenCode.left_parenthesis},
+			{')', TokenCode.right_parenthesis},
+			{'[', TokenCode.left_bracket},
+			{']', TokenCode.right_bracket},
+			{'*', TokenCode.factor_op},
+			{'%', TokenCode.factor_op},
+			{'+', TokenCode.term_op}, 
+			{'-', TokenCode.term_op},
+			{';', TokenCode.semicolon},
+		};
+
+	public static bool Contains(char elem) =>
+		separatorCodes.Keys.Contains(elem);
+
+	public static TokenCode Code(char ch) =>
+		separatorCodes[ch];
+}
+
+public class Token {
+	private Position position {get;}
+	private TokenCode tokenCode {get;}
+	private object? value {get; set;}
+
+	public Position Position() => this.position;
+	public TokenCode Code() => this.tokenCode;
+	public object Value() => this.value is null ? "" : this.value;
+
+	public void Value(object val) => this.value = val;
+
+	// Construction
+	public Token(Position pos, TokenCode code, object? val = null) {
+		this.position = pos;
+		this.tokenCode = code;
+		if(val != null)
+			this.value = val;
 	}
 
-	public static class SeparatorList {
-		private static Dictionary<char, TokenCode> separatorCodes =
-			new Dictionary<char, TokenCode>() {
-				{',', TokenCode.comma},
-				{'(', TokenCode.left_parenthesis},
-				{')', TokenCode.right_parenthesis},
-				{'[', TokenCode.left_bracket},
-				{']', TokenCode.right_bracket},
-				{'*', TokenCode.factor_op},
-				{'%', TokenCode.factor_op},
-				{'+', TokenCode.term_op}, 
-				{'-', TokenCode.term_op},
-				{';', TokenCode.semicolon},
-			};
-
-		public static bool Contains(char elem) =>
-			separatorCodes.Keys.Contains(elem);
-
-		public static TokenCode Code(char ch) =>
-			separatorCodes[ch];
-	}
-
-	public class Token {
-		private Position position {get;}
-		private TokenCode tokenCode {get;}
-		private object value {get; set;}
-
-		public Position Position() => this.position;
-		public TokenCode Code() => this.tokenCode;
-		public object Value() => this.value;
-
-		public void Value(object val) => this.value = val;
-
-		// Construction
-		public Token(Position pos, TokenCode code, object? val = null) {
-			this.position = pos;
-			this.tokenCode = code;
-			this.value = "";
-			if(val != null)
-				this.value = val;
-		}
-
-		public void PrintInfo() {
-			Console.Write($"{this.position.ToString()}\t{this.tokenCode}\t\t");
-			if(!(this.value is null))
-				Console.Write($"Value: {this.value}");
-			Console.Write("\n");
-		}
+	public void PrintInfo() {
+		Console.Write("{0,-10} {1,-25}", this.position.ToString(), this.tokenCode);
+		if(this.value != null)
+			Console.Write($"Value: {this.value}");
+		Console.Write("\n");
 	}
 }
